% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_summarize.R
\name{filter_overlaps}
\alias{filter_overlaps}
\title{Filter overlaps}
\usage{
filter_overlaps(
  distance = 3,
  detail_file,
  out_file = file.path(dirname(detail_file), paste0(sub(".csv", "",
    basename(detail_file)), "_filter_", distance, ".csv")),
  validate = TRUE,
  chopoff_path = install_CHOPOFF()
)
}
\arguments{
\item{distance}{What is the distance for overlap filtering of off-targets. (default: 3)}

\item{detail_file}{Path to the file where database is stored (output of search_index function, .csv file)}

\item{out_file}{File path to the file where summarized output should be generated.}

\item{validate}{TRUE, if false, do not check that CHOPOFF path is valid}

\item{chopoff_path}{PATH to CHOPOFF, default install_chopoff()}
}
\value{
path to csv output file of filtered guides
}
\description{
Filter overlaps
}
\examples{
name <- "CAS9"
genome <- system.file("extdata/sample_genome", "semirandom.fa", package = "crisprCHOPOFF")
## Note: a fasta index ".fai" file must exist in directory of genome.
# You can make it with:
#if (!file.exists(paste0(genome, ".fai"))) {
#Rsamtools::indexFa(genome)
#}
out_dir_index <- file.path(tempdir(), "CHOPOFF_sample_genome")
build_index(name, genome, out_dir_index, validate = FALSE)

# Now search some guides
guides <- system.file("extdata/sample_genome", "guides.txt", package = "crisprCHOPOFF")
# Quick preview in guides:
guide_candidates <- read.table(guides, col.names = "guides")
unique(nchar(unlist(guide_candidates))) # Unique lengths of guides
guide_hits <- search_index(guides, out_dir_index, validate = FALSE)

filter_overlaps(distance = 3, guide_hits)
}
