% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index.R
\name{install_CHOPOFF}
\alias{install_CHOPOFF}
\title{Install backend utilities for CHOPOFF}
\usage{
install_CHOPOFF(
  script = system.file("bash_script", "install_julia_and_CHOPOFF.sh", package =
    "crisprCHOPOFF")
)
}
\arguments{
\item{script}{the bash script to run to install the backend}
}
\value{
The path to installed CHOPOFF.jl binary, can also be retrieved with Sys.getenv("CHOPOFF").
}
\description{
Will install from source:\cr
\itemize{
\item Julia 1.8.5\cr
\item CHOPOFF.jl
}
}
