% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index.R
\name{search_index}
\alias{search_index}
\title{Search guides in CHOPOFF index}
\usage{
search_index(
  guides,
  index_dir,
  out_file = file.path(index_dir, paste0(algorithm, "_", distance, ".csv")),
  algorithm = "prefixHashDB",
  distance = 3,
  validate = TRUE,
  chopoff_path = Sys.getenv("CHOPOFF")
)
}
\arguments{
\item{guides}{path to txt file of guides, 1 per line of correct length}

\item{algorithm}{default: prefixHashDB, alternatives: TODO: list all here!}

\item{distance}{""}

\item{validate}{TRUE, if false, do not check that CHOPOFF path is valid}

\item{chopoff_path}{PATH to CHOPOFF, default "CHOPOFF"}

\item{out_dir}{""}
}
\value{
path to csv output file of guides
}
\description{
Search guides in CHOPOFF index
}
\examples{
name <- "CAS9"
genome <- system.file("extdata/sample_genome", "semirandom.fa", package = "crisprCHOPOFF")
## Note: a fasta index ".fai" file must exist in directory of genome.
# You can make it with:
#if (!file.exists(paste0(genome, ".fai"))) {
#Rsamtools::indexFa(genome)
#}
out_dir_index <- file.path(tempdir(), "CHOPOFF_sample_genome")
build_index(name, genome, out_dir_index, validate = FALSE)

# Now search some guides
guides <- system.file("extdata/sample_genome", "guides.txt", package = "crisprCHOPOFF")
guide_hits <- search_index(guides, out_dir_index, validate = FALSE)
guide_hits_table <- read.table(guide_hits, sep = ",", header = TRUE)
# use data.table::fread for reading in large list
# Subset to 0 distance hits
dist0 <- guide_hits_table[guide_hits_table$distance == 0,]
head(dist0)
# Which chromosomes is a specific guide found on with 0 distance hits?
unique(dist0[dist0$guide == "TCCGGCCTGGTTATCGAAGG",]$chromosome) # 2 chromosomes
}
