% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index.R
\name{build_index}
\alias{build_index}
\title{Build CHOPOFF index}
\usage{
build_index(
  name,
  genome,
  out_dir,
  algorithm = "prefixHashDB",
  distance = 3,
  preset = "Cas9",
  hash_length = 16,
  ambig_max = 0,
  strands = c("+", "-"),
  fwd_motif = "NNNNNNNNNNNNNNNNNNNNXXX",
  fwd_pam = "XXXXXXXXXXXXXXXXXXXXNGG",
  extend3prime = FALSE,
  validate = TRUE,
  chopoff_path = install_CHOPOFF()
)
}
\arguments{
\item{name}{How will you shortly name this database?}

\item{genome}{Path to the genome, either .fa or .2bit. Must have a fasta
index in directory, named \code{paste0(genome, ".fai")}}

\item{out_dir}{output directory where the database will be stored}

\item{algorithm}{default: prefixHashDB, alternatives: TODO: list all here!}

\item{distance}{Within what distance can we search for off-targets? (default: 3)}

\item{preset}{The preset for parameters, default "Cas9". Alternatives:
Cas12 and NULL (user custom settings)}

\item{hash_length}{""}

\item{ambig_max}{How many ambiguous bases are allowed inside the guide? (default: 0)}

\item{strands}{c("+", "-"), search both 5' (+) and 3' (-) strands}

\item{fwd_motif}{How many ambiguous bases are allowed inside the guide? (default: 0)}

\item{fwd_pam}{Motif in 5'-3' that will be matched on the reference (without the X). For example for Cas9 it is 20*N + NGG: XXXXXXXXXXXXXXXXXXXXNGG}

\item{extend3prime}{Defines how off-targets will be aligned to the
guides and where extra nucleotides will be
added for alignment within distance. Whether
to extend in the 5' and 3' direction. Default
is Cas9 with extend3 = false.}

\item{validate}{TRUE, if false, do not check that CHOPOFF path is valid}

\item{chopoff_path}{PATH to CHOPOFF, default install_CHOPOFF()}
}
\value{
invisible(NULL)
}
\description{
Build the genome index for CHOPOFF, using either a preset (Cas9 or Cas12),
or user defined parameters.
}
\examples{
name <- "CAS9"
genome <- system.file("extdata/sample_genome", "semirandom.fa", package = "crisprCHOPOFF")
## Note: a fasta index ".fai" file must exist in directory of genome.
# You can make it with:
#if (!file.exists(paste0(genome, ".fai"))) {
#Rsamtools::indexFa(genome)
#}
out_dir_index <- file.path(tempdir(), "CHOPOFF_sample_genome")
build_index(name, genome, out_dir_index, validate = FALSE)

}
