% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index.R
\name{build_index}
\alias{build_index}
\title{Build CHOPOFF index}
\usage{
build_index(
  name,
  genome,
  out_dir,
  algorithm = "prefixHashDB",
  distance = 3,
  motif = "Cas9",
  hash_length = 16,
  ambig_max = 0,
  strands = c("+", "-"),
  fwd_motif = "NNNNNNNNNNNNNNNNNNNNXXX",
  fwd_pam = "XXXXXXXXXXXXXXXXXXXXNGG",
  extend3prime = FALSE,
  validate = TRUE,
  chopoff_path = install_CHOPOFF()
)
}
\arguments{
\item{name}{How will you shortly name this database?}

\item{genome}{Path to the genome, either .fa or .2bit. Must have a fasta
index in directory, named \code{paste0(genome, ".fai")}}

\item{out_dir}{""}

\item{algorithm}{default: prefixHashDB, alternatives: TODO: list all here!}

\item{distance}{""}

\item{motif}{The preset for parameters, default "Cas9". Alternatives:
Cas12 and NULL (user custom settings)}

\item{hash_length}{""}

\item{ambig_max}{""}

\item{strands}{c("+", "-"), search both 5' (+) and 3' (-) strands}

\item{fwd_motif}{""}

\item{fwd_pam}{""}

\item{extend3prime}{""}

\item{validate}{TRUE, if false, do not check that CHOPOFF path is valid}

\item{chopoff_path}{PATH to CHOPOFF, default install_CHOPOFF()}
}
\value{
invisible(NULL)
}
\description{
Build CHOPOFF index
}
\examples{
name <- "CAS9"
genome <- system.file("extdata/sample_genome", "semirandom.fa", package = "crisprCHOPOFF")
## Note: a fasta index ".fai" file must exist in directory of genome.
# You can make it with:
#if (!file.exists(paste0(genome, ".fai"))) {
#Rsamtools::indexFa(genome)
#}
out_dir_index <- file.path(tempdir(), "CHOPOFF_sample_genome")
build_index(name, genome, out_dir_index, validate = FALSE)

}
